+++ packages/api-background-tasks-os/package.json
@@ -15,7 +15,6 @@
   "license": "MIT",
   "dependencies": {
     "@webiny/api-elasticsearch-tasks": "0.0.0",
-    "@webiny/api-headless-cms-es-tasks": "0.0.0",
     "@webiny/plugins": "0.0.0"
   },
   "devDependencies": {
