+++ packages/api-headless-cms-es-tasks/package.json
@@ -17,13 +17,13 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@webiny/api": "0.0.0",
     "@webiny/api-core": "0.0.0",
     "@webiny/api-headless-cms": "0.0.0",
     "@webiny/api-headless-cms-ddb-es": "0.0.0",
     "@webiny/api-opensearch": "0.0.0",
     "@webiny/background-tasks": "0.0.0",
     "@webiny/error": "0.0.0",
+    "@webiny/feature": "0.0.0",
     "@webiny/utils": "0.0.0"
   },
   "devDependencies": {
