+++ packages/api-headless-cms-es-tasks/tsconfig.json
@@ -2,13 +2,13 @@
   "extends": "../../tsconfig.json",
   "include": ["src", "__tests__"],
   "references": [
-    { "path": "../api" },
     { "path": "../api-core" },
     { "path": "../api-headless-cms" },
     { "path": "../api-headless-cms-ddb-es" },
     { "path": "../api-opensearch" },
     { "path": "../background-tasks" },
     { "path": "../error" },
+    { "path": "../feature" },
     { "path": "../utils" },
     { "path": "../handler-db" },
     { "path": "../handler-graphql" },
@@ -21,8 +21,6 @@
     "paths": {
       "~/*": ["./src/*"],
       "~tests/*": ["./__tests__/*"],
-      "@webiny/api/*": ["../api/src/*"],
-      "@webiny/api": ["../api/src"],
       "@webiny/api-core/*": ["../api-core/src/*"],
       "@webiny/api-core": ["../api-core/src"],
       "@webiny/api-headless-cms/*": ["../api-headless-cms/src/*"],
@@ -35,6 +33,8 @@
       "@webiny/background-tasks": ["../background-tasks/src"],
       "@webiny/error/*": ["../error/src/*"],
       "@webiny/error": ["../error/src"],
+      "@webiny/feature/*": ["../feature/src/*"],
+      "@webiny/feature": ["../feature/src"],
       "@webiny/utils/*": ["../utils/src/*"],
       "@webiny/utils": ["../utils/src"],
       "@webiny/handler-db/*": ["../handler-db/src/*"],
